Добавляет разделение кода по модулям main, data и utilities
